+++ src/rules/no-restricted-paths.js
@@ -68,6 +68,19 @@ module.exports = {
       });
     }
 
+    const zoneExceptions = matchingZones.map((zone) => {
+      const exceptionPaths = zone.except || [];
+      const absoluteFrom = path.resolve(basePath, zone.from);
+      const absoluteExceptionPaths = exceptionPaths.map((exceptionPath) => path.resolve(absoluteFrom, exceptionPath));
+      const hasValidExceptionPaths = absoluteExceptionPaths
+        .every((absoluteExceptionPath) => isValidExceptionPath(absoluteFrom, absoluteExceptionPath));
+
+      return {
+        absoluteExceptionPaths,
+        hasValidExceptionPaths,
+      };
+    });
+
     function checkForRestrictedImportPath(importPath, node) {
       const absoluteImportPath = resolve(importPath, context);
 
@@ -75,19 +88,14 @@ module.exports = {
         return;
       }
 
-      matchingZones.forEach((zone) => {
-        const exceptionPaths = zone.except || [];
+      matchingZones.forEach((zone, index) => {
         const absoluteFrom = path.resolve(basePath, zone.from);
 
         if (!containsPath(absoluteImportPath, absoluteFrom)) {
           return;
         }
 
-        const absoluteExceptionPaths = exceptionPaths.map((exceptionPath) =>
-          path.resolve(absoluteFrom, exceptionPath)
-        );
-        const hasValidExceptionPaths = absoluteExceptionPaths
-          .every((absoluteExceptionPath) => isValidExceptionPath(absoluteFrom, absoluteExceptionPath));
+        const { hasValidExceptionPaths, absoluteExceptionPaths } = zoneExceptions[index];
 
         if (!hasValidExceptionPaths) {
           reportInvalidExceptionPath(node);